* Paperclip compatibility need to be required manually * Test matchers need to be required manually * Framework integration needs to be required manually * Drop support for sinatra before 1.3.0
