+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ## Removed
 * Remove support for Merb (@seuros [#2566](https://github.com/carrierwaveuploader/carrierwave/pull/2566))
+* Paperclip compatibility need to be required manually
+* Test matchers need to be required manually
+* Framework integration needs to be required manually
+* Drop support for sinatra before 1.3.0
 
 ## 2.2.2 - 2021-05-28
 ### Fixed
