+++ lib/carrierwave/frameworks/padrino.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+gem "padrino-core", "~> 0.10.5"
+CarrierWave.root = File.join(PADRINO_ROOT, "public")
